Shut up warning. We’ll just have to live with a 4 billion lines limit on scripts for now.

+++ macosx/LEORemoteDebugger.c
@@ -187,7 +187,7 @@ void LEORemoteDebuggerUpdateState( struct LEOContext* inContext )
 	{
 		LEOScript*	script = LEOContextPeekCurrentScript(inContext);
 		uint16_t	fileID = script->parseErrors[inContext->currentInstruction->param2].fileID;
-		uint32_t	lineNumber = script->parseErrors[inContext->currentInstruction->param2].errorLine;
+		uint32_t	lineNumber = (uint32_t)script->parseErrors[inContext->currentInstruction->param2].errorLine;
 		
 		actuallyWritten = write( gLEORemoteDebuggerSocketFD, "LINE", 4 );
 		dataLen = sizeof(fileID) + sizeof(lineNumber);
